% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-date-range.R
\name{date_range_presets_vec}
\alias{date_range_presets_vec}
\title{Date Range Presets}
\usage{
date_range_presets_vec(max_date = NULL)
}
\arguments{
\item{max_date}{The max date for date range to return}
}
\description{
This function returns a vector of preset date ranges like Last N Weeks.
}
