% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-csv.R
\name{download_csv}
\alias{download_csv}
\alias{download_csv_ui}
\title{Download Data as CSV}
\usage{
download_csv(input, output, session, dataset, filename, ...)

download_csv_ui(id, ...)
}
\arguments{
\item{input}{standard \code{shiny} boilerplate}

\item{output}{standard \code{shiny} boilerplate}

\item{session}{standard \code{shiny} boilerplate}

\item{dataset}{a data frame, or a function/reactive that returns a data frame}

\item{filename}{a string, or a function/reactive that returns a string}

\item{...}{additional parameters to pass to \code{\link[utils]{write.csv}}}

\item{id}{a string indicating the id to use the module with.}
}
\description{
A shiny module that adds a download button to download data as a CSV.
}
\examples{
\dontrun{
  shinybones::preview_module(download_csv,
    dataset = mtcars,
    filename = function(){
      paste0('mtcars-', format(Sys.time(), "\%Y-\%m-\%d-\%H-\%M-\%S"), '.csv')
    }
  )
}
}
