% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-panel-metric.R
\name{get_dimension_tabs}
\alias{get_dimension_tabs}
\title{Get dimension tabs}
\usage{
get_dimension_tabs(metric, hidden_dimensions = NULL)
}
\arguments{
\item{metric}{a \code{tbl_metric} object}

\item{hidden_dimensions}{a vector of dimensions that should be hidden}
}
\description{
Get dimension tabs
}
\examples{
library(dplyr)
\dontrun{
metrics_condensed <- datacampr::dc_s3_read(
 'metrics_condensed.rds'
)$metrics_condensed
metrics_condensed$finance_churn_pct_b2c_churn \%>\%
  get_dimension_tabs()
}
}
