% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plots.R
\name{get_plot_type}
\alias{get_plot_type}
\title{Get default plot_type}
\usage{
get_plot_type(metric_id, plot_type = NULL)
}
\arguments{
\item{metric_id}{metric id}

\item{plot_type}{plot type}
}
\description{
Get default plot_type
}
\examples{
\dontrun{
get_plot_type('pct_avg_rating')
get_plot_type('nb_started')
get_plot_type('nb_started', 'line')
}
}
