% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_value}
\alias{get_value}
\title{S3 method to get the value of an object}
\usage{
get_value(x, ...)
}
\arguments{
\item{x}{an object whose value is to be retrieved}

\item{...}{additional parameters passed on to the method.}
}
\description{
S3 method to get the value of an object
}
