% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-date-range.R
\name{input_date_range}
\alias{input_date_range}
\alias{input_date_range_ui}
\title{Create date range input with preset ranges}
\usage{
input_date_range(input, output, session, date_range = c(Sys.Date() - 365,
  Sys.Date()), selected_date_range_preset = "Last Year", ...)

input_date_range_ui(id, ...)
}
\arguments{
\item{input}{standard \code{shiny} boilerplate}

\item{output}{standard \code{shiny} boilerplate}

\item{session}{standard \code{shiny} boilerplate}

\item{date_range}{a range of dates}

\item{selected_date_range_preset}{selected date range preset}

\item{...}{additional parameters passed to the module}

\item{id}{a string indicating the id to call the module with}
}
\value{
A reactive vector of the selected date range
}
\description{
This is a shiny module to create a date range input with presets for the
Last 1 week, 2 weeks, 6 months etc.
}
\examples{
library(shiny)
\dontrun{
 shinybones::preview_module(input_date_range, use_box = TRUE)
 test_date_range <- function(input, output, session, ...){
   ns <- session$ns
   date_range_input <-  callModule(input_date_range, 'date_range')
   output$date_range_text <- renderText({
     paste(date_range_input(), collapse = " - ")
   })
 }
 test_date_range_ui <- function(id, ...){
   ns <- shiny::NS(id)
   shinydashboard::box(
     width = 12,
     title = 'Date Range Input',
     input_date_range_ui(ns('date_range')),
     column(12, textOutput(ns('date_range_text')))
   )
 }
 shinybones::preview_module(test_date_range)
}
}
