% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-select-period.R
\name{input_select_period}
\alias{input_select_period}
\title{Create a picker input to select aggregation period}
\usage{
input_select_period(inputId, selected_period = NULL, periods = "All",
  label = "aggregated_by", select_func = shiny::selectInput, ...)
}
\arguments{
\item{inputId}{the input slot that will be used to access the value.}

\item{selected_period}{a string indicating selected period}

\item{periods}{a named vector of periods.}

\item{label}{a string to display as label.}

\item{select_func}{a select input function. Either
\code{\link[shiny]{selectInput}} or \code{\link[shinyWidgets]{pickerInput}}}

\item{...}{additional parameters passed on to \code{select_func}}
}
\description{
Create a picker input to select aggregation period
}
\examples{
\dontrun{
input_select_period('period') \%>\%
  shinybones::preview_component()
input_select_period('period', selected_period = 'week') \%>\%
  shinybones::preview_component()
}
}
