% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-toggle-pct.R
\name{input_toggle_pct}
\alias{input_toggle_pct}
\title{Radio buttons to toggle between number and percentage.}
\usage{
input_toggle_pct(inputId, label = "as", size = "sm", ...)
}
\arguments{
\item{inputId}{the input slot that will be used to access the value.}

\item{label}{label to display alongside the buttons}

\item{size}{size of the radio buttons ('sm', 'xs', or 'lg')}

\item{...}{additional parameters to pass to
\code{\link[shinyWidgets]{radioGroupButtons}}}
}
\description{
Radio buttons to toggle between number and percentage.
}
\examples{
\dontrun{
  input_toggle_pct('show_pct') \%>\%
    shinybones::preview_component()
}
}
