% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{make_toast_position}
\alias{make_toast_position}
\title{Translate toast postion to CSS}
\usage{
make_toast_position(position)
}
\arguments{
\item{position}{Position string.}
}
\description{
Translate toast postion to CSS
}
\keyword{internal}
