% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-panel.R
\name{metric_panel}
\alias{metric_panel}
\alias{metric_panel_ui}
\title{Display a metric in a panel}
\usage{
metric_panel(input, output, session, metric, plot_type = NULL,
  plot_fun = NULL, plot_post_process = NULL,
  orientation = "vertical", selected_date_range_preset = "All Time",
  selected_period = "Week", height = 400, div_bottom_left = NULL,
  hidden_dimensions = NULL, ...)

metric_panel_ui(id, ...)
}
\arguments{
\item{input}{standard \code{shiny} boilerplate}

\item{output}{standard \code{shiny} boilerplate}

\item{session}{standard \code{shiny} boilerplate}

\item{metric}{A metric to display. It should be a list with data and title}

\item{plot_type}{Either "line" or "bar": if neither is given, it guesses
"line" for \code{pct}, \code{avg}, \code{min}, \code{median}, and \code{max}
metrics, and "bar" otherwise. Ignored if \code{plot_fun} is provided.}

\item{plot_fun}{A plotting function that takes two arguments - data and
dimension}

\item{plot_post_process}{a function to post-process the default plot}

\item{orientation}{a string indicating orientation (vertical or horizontal)}

\item{selected_date_range_preset}{Default date range preset to use. Use
\code{date_range_presets_vec()} to look up possible presets}

\item{selected_period}{Default period to display the data}

\item{height}{height of the panel}

\item{div_bottom_left}{html to display on bottom left of the panel}

\item{hidden_dimensions}{A vector of dimension names that should not be
displayed as tabs.}

\item{...}{Additional parameters passed to the server}

\item{id}{a string indicating the id to call the module with}
}
\description{
Display a metric in a panel
}
\examples{
library(dplyr)
\dontrun{
 metrics_condensed <- datacampr::dc_s3_read(
   "metrics_condensed.rds"
 )$metrics_condensed
 preview_metric(
   metric = metrics_condensed$finance_forecasts_usd_arr_total,
   plot_type = 'line'
 )
 preview_metric(metrics_condensed$finance_cash_usd_cash_in)
}
}
