% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-panel-footer.R
\name{metric_panel_footer}
\alias{metric_panel_footer}
\alias{metric_panel_footer_ui}
\title{Metric Panel Footer}
\usage{
metric_panel_footer(input, output, session, metric,
  date_range = range(metric$date),
  selected_date_range_preset = "Last Year", ...)

metric_panel_footer_ui(id, selected_period = NULL, periods, ...)
}
\arguments{
\item{input}{standard \code{shiny} boilerplate}

\item{output}{standard \code{shiny} boilerplate}

\item{session}{standard \code{shiny} boilerplate}

\item{metric}{a \code{tbl_metric} object}

\item{date_range}{a range of dates}

\item{selected_date_range_preset}{selected date range preset}

\item{...}{Additional parameters to pass to module}

\item{id}{a string indicating the id to call the module with}

\item{selected_period}{a string indicating selected period}

\item{periods}{a named vector of periods.}
}
\description{
This is a component module to ...
}
\examples{
\dontrun{
shinybones::preview_module(metric_panel_footer,
  selected_date_range_preset = 'Last Week',
  selected_period = 'month'
)
}
\dontrun{
library(dplyr)
metrics <- datacampr::dc_s3_read('metrics_condensed.rds')$metrics_condensed
preview_metric(metrics$content_courses_avg_rating_wtd)
}
}
