% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{metrics Produce metrics for shiny app}
\usage{
metrics(
  app = "FINDCov19Policy",
  app_name = "FIND Sars-CoV-2 Policy dashboard",
  server = "shinyapps.io",
  from = "2022-09-01 0:00:00",
  until = "2022-12-12 0:00:00"
)
}
\arguments{
\item{app}{The name of the app on the shinyapps.io server}

\item{app_name}{The name of the the application, to be used for plots and titles}

\item{server}{The server where the shiny app is hosted}

\item{from}{The date and time from when the metrics should be measured from}

\item{until}{The date and time until when the metrics should be measured from}
}
\value{
A data frame with number of connections per day for the given shiny application
}
\description{
metrics Produce metrics for shiny app
}
\examples{
app <- c("FINDCov19Tracker")
app_name <- c("FIND Sars-CoV-2 Test tracker")
server <- rep("shinyapps.io", length(app))
from <- rep("2023-09-01 0:00:00", length(app))
until <- rep("2023-12-12 0:00:00", length(app))

metrics_df <- metrics(app = app, app_name = app_name, 
                                    server = server, from = from, until = until)

}
