% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics_df}
\alias{metrics_df}
\title{Metrics data frame}
\format{
A data frame with 103 rows and 3 variables:
\describe{
\item{\code{date}}{double date }
\item{\code{key}}{character application name}
\item{\code{value}}{double number of connections}
}
}
\usage{
metrics_df
}
\description{
Data frame containing metrics, i.e. number of connections for a shiny application
}
\keyword{datasets}
