% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_metric_condensed}
\alias{plot_metric_condensed}
\title{Plot a condensed metric}
\usage{
plot_metric_condensed(metric, dimension = "all", plot_type = NULL,
  quietly = TRUE, ...)
}
\arguments{
\item{metric}{a \code{tbl_metric} object}

\item{dimension}{a string indicating the dimension to visualize}

\item{plot_type}{a string indicating type of plot (line/bar)}

\item{quietly}{a boolean indicating if messages should be suppressed}

\item{...}{additional parameters passed}
}
\description{
Plot a condensed metric
}
\examples{
\dontrun{
  library(dplyr)
  metrics_condensed <- datacampr::dc_s3_read(
    'metrics_condensed.rds'
  )$metrics_condensed
  metric \%>\%
    filter(period == 'week') \%>\%
    filter(date >= Sys.Date() - 365) \%>\%
    plot_metric_condensed()

 metrics_condensed$product_time_median_time_spent \%>\%
   filter(period == 'week') \%>\%
   filter(date >= Sys.Date() - 365) \%>\%
   plot_metric_condensed_line(dimension = 'subscription_type')

 metrics_condensed$content_courses_avg_rating_wtd \%>\%
   filter(period == 'rolling_28d') \%>\%
   filter(date >= Sys.Date() - 365) \%>\%
   plot_metric_condensed()

 metrics_condensed$content_courses_avg_rating_wtd \%>\%
   filter(period == 'rolling_28d') \%>\%
   filter(date >= Sys.Date() - 365) \%>\%
   plot_metric_condensed('ds_track')

 flights_nyc_avg_arr_delay \%>\%
   filter(period == 'week') \%>\%
   plot_metric_condensed(plot_type = 'bar')

 flights_nyc_avg_arr_delay \%>\%
   filter(period == 'week') \%>\%
   plot_metric_condensed(plot_type = 'bar', dimension = 'origin')

 flights_nyc_avg_arr_delay \%>\%
   filter(period == 'week') \%>\%
   plot_metric_condensed(plot_type = 'line')

 flights_nyc_avg_arr_delay \%>\%
   filter(period == 'week') \%>\%
   plot_metric_condensed(plot_type = 'line', dimension = 'origin')
}
}
