% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_metrics}
\alias{plot_metrics}
\title{plot_metrics Produce metrics for shiny app}
\usage{
plot_metrics(
  df = metrics_df,
  weekly = TRUE,
  app_name = "FIND Sars-CoV-2 Test tracker"
)
}
\arguments{
\item{df}{A data frame with number of connections per day for the given shiny application, produced by metrics function}

\item{weekly}{Logical variable, TRUE if plot should be showcasing number of connections by week, FALSE for number of connections per day}

\item{app_name}{The name of the the application, to be used for plots and titles}
}
\value{
A plot with number of connections per day for the given shiny application
}
\description{
plot_metrics Produce metrics for shiny app
}
\examples{
p <- plot_metrics(df = metrics_df, weekly = TRUE, app_name = "FIND Sars-CoV-2 Test tracker")
print(p)

}
