% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{Shinymetrics}
\alias{Shinymetrics}
\title{Shinymetrics}
\description{
Shinymetrics

Shinymetrics
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{token}}{Your \href{https://shinymetrics.com}{shinymetrics} token.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Shinymetrics-new}{\code{Shinymetrics$new()}}
\item \href{#method-Shinymetrics-track_recommended}{\code{Shinymetrics$track_recommended()}}
\item \href{#method-Shinymetrics-track_connected}{\code{Shinymetrics$track_connected()}}
\item \href{#method-Shinymetrics-track_disconnected}{\code{Shinymetrics$track_disconnected()}}
\item \href{#method-Shinymetrics-track_session_initialialized}{\code{Shinymetrics$track_session_initialialized()}}
\item \href{#method-Shinymetrics-track_busy_idle}{\code{Shinymetrics$track_busy_idle()}}
\item \href{#method-Shinymetrics-track_input}{\code{Shinymetrics$track_input()}}
\item \href{#method-Shinymetrics-track_message}{\code{Shinymetrics$track_message()}}
\item \href{#method-Shinymetrics-track_conditional}{\code{Shinymetrics$track_conditional()}}
\item \href{#method-Shinymetrics-track_bound}{\code{Shinymetrics$track_bound()}}
\item \href{#method-Shinymetrics-track_unbound}{\code{Shinymetrics$track_unbound()}}
\item \href{#method-Shinymetrics-track_output}{\code{Shinymetrics$track_output()}}
\item \href{#method-Shinymetrics-track_error}{\code{Shinymetrics$track_error()}}
\item \href{#method-Shinymetrics-track_output_invalidated}{\code{Shinymetrics$track_output_invalidated()}}
\item \href{#method-Shinymetrics-track_recalculating}{\code{Shinymetrics$track_recalculating()}}
\item \href{#method-Shinymetrics-track_recalculated}{\code{Shinymetrics$track_recalculated()}}
\item \href{#method-Shinymetrics-track_visual_change}{\code{Shinymetrics$track_visual_change()}}
\item \href{#method-Shinymetrics-track_update_input}{\code{Shinymetrics$track_update_input()}}
\item \href{#method-Shinymetrics-track_page_view}{\code{Shinymetrics$track_page_view()}}
\item \href{#method-Shinymetrics-track_file_download}{\code{Shinymetrics$track_file_download()}}
\item \href{#method-Shinymetrics-track_custom}{\code{Shinymetrics$track_custom()}}
\item \href{#method-Shinymetrics-include}{\code{Shinymetrics$include()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-new"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$new(
  token = Sys.getenv("SHINYMETRICS_TOKEN"),
  prod = getOption("SHINYMETRICS_PROD", FALSE),
  ignore_global = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{Your application token.}

\item{\code{prod}}{Whether the events are tracked for production.}

\item{\code{ignore_global}}{Whether to ignore "global" events,
these are events that target the entire HTML document,
they tend to add noise.
They mainly are events sent by HTMLwidgetsand the likes.
It is advised to ignore them.

These allow not polluting your dashboard with test data.
It is \code{FALSE} by default.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialize tracker
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_recommended"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_recommended}{}}}
\subsection{Method \code{track_recommended()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_recommended()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Recommended events to track.

Tracks the following:
\itemize{
\item \code{input}
\item \code{output}
\item \code{error}
\item \code{file_download}
\item \code{page_view}
\item \code{busy_idle}
}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_connected"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_connected}{}}}
\subsection{Method \code{track_connected()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_connected(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "connected" event fired when
Shiny's server connects.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_disconnected"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_disconnected}{}}}
\subsection{Method \code{track_disconnected()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_disconnected(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "disconnected" event fired when
Shiny's server disconnects.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_session_initialialized"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_session_initialialized}{}}}
\subsection{Method \code{track_session_initialialized()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_session_initialialized(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "sessioninitialized" event fired when
Shiny's session is initialized.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_busy_idle"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_busy_idle}{}}}
\subsection{Method \code{track_busy_idle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_busy_idle(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "busy" and "idle" event fired when the server is performing
calculations and stops.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_input"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_input}{}}}
\subsection{Method \code{track_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_input(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "inputchanged" event fired when an
input value changes.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_message"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_message}{}}}
\subsection{Method \code{track_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_message(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "message" event triggered when any messages
are received from the server.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_conditional"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_conditional}{}}}
\subsection{Method \code{track_conditional()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_conditional(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "conditional" triggered when \code{shiny::conditionalPanel}
are updated.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_bound"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_bound}{}}}
\subsection{Method \code{track_bound()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_bound(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "bound" triggered when an input or output
is bound with the shiny server.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_unbound"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_unbound}{}}}
\subsection{Method \code{track_unbound()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_unbound(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "unbound" triggered when an input or output
is no longer bound with the shiny server.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_output"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_output}{}}}
\subsection{Method \code{track_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_output(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track the "value" event which is triggered when
an output receives a value from the server
(when an output is updated).
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_error"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_error}{}}}
\subsection{Method \code{track_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_error(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track the "error" event, triggered when an
error is propagated to an output
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_output_invalidated"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_output_invalidated}{}}}
\subsection{Method \code{track_output_invalidated()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_output_invalidated(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "outputinvalidated" event triggered,
when an output’s value is invalidated on the server.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_recalculating"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_recalculating}{}}}
\subsection{Method \code{track_recalculating()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_recalculating(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "recalculating" triggered before an output
value is recalculated.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_recalculated"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_recalculated}{}}}
\subsection{Method \code{track_recalculated()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_recalculated(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "recalculating" triggered after an output
value has been recalculated.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_visual_change"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_visual_change}{}}}
\subsection{Method \code{track_visual_change()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_visual_change(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "visualchange" triggered when an output is
resized, hidden, or shown.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_update_input"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_update_input}{}}}
\subsection{Method \code{track_update_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_update_input(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "updateinput" event is triggered when an input
is updated from the server, e.g., when you call \code{\link[shiny:updateTextInput]{shiny::updateTextInput()}}
in R to update the label or value of a text input.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_page_view"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_page_view}{}}}
\subsection{Method \code{track_page_view()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_page_view(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track when a page is loaded.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_file_download"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_file_download}{}}}
\subsection{Method \code{track_file_download()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_file_download(ignore = c(), only = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ignore}}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{\code{only}}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track "filedownload" event triggered when a file is downloaded
via the \code{\link[shiny:downloadButton]{shiny::downloadButton()}}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-track_custom"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-track_custom}{}}}
\subsection{Method \code{track_custom()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$track_custom(
  type,
  id = NULL,
  value = NULL,
  session = shiny::getDefaultReactiveDomain()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Type of event that is pushed.}

\item{\code{id}}{A unique identifier for your custom event, used
to identify specific events of a certain \code{type}.}

\item{\code{value}}{Value of the event, this is converted to a character.}

\item{\code{session}}{A valid Shiny session.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Track custom events.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Shinymetrics-include"></a>}}
\if{latex}{\out{\hypertarget{method-Shinymetrics-include}{}}}
\subsection{Method \code{include()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Shinymetrics$include()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Include the tracking code in your UI.
}

}
}
