% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{shinymetrics_custom_event}
\alias{shinymetrics_custom_event}
\title{Shinymetrics Custom Event}
\usage{
shinymetrics_custom_event(
  type,
  id = NULL,
  value = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{type}{Type of event that is pushed.}

\item{id}{A unique identifier for your custom event, used
to identify specific events of a certain \code{type}.}

\item{value}{Value of the event, this is converted to a character.}

\item{session}{A valid Shiny session.}
}
\description{
Push \strong{custom events} from the server.
Requires running \code{\link[=shinymetrics_server]{shinymetrics_server()}} prior.
}
\details{
It will not work for standard events.

Standard events:
\itemize{
\item \code{connected}
\item \code{disconnected}
\item \code{disconnected}
\item \code{sessioninitialised}
\item \code{busy}
\item \code{idle}
\item \code{input}
\item \code{output}
\item \code{message}
\item \code{conditional}
\item \code{bound}
\item \code{unbound}
\item \code{error}
\item \code{outputinvalidated}
\item \code{recalculating}
\item \code{recalculated}
\item \code{visualchange}
\item \code{updateinput}
\item \code{pageview}
\item \code{filedownload}
}

These must be custom events of your choosing, e.g.:
a custom event when registering or logging in a user.
}
