% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles.R
\name{title_with_modal}
\alias{title_with_modal}
\alias{title_with_tooltip}
\alias{title_with_popover}
\title{Title with help text in an informative modal, popover or tooltip}
\usage{
title_with_modal(title, help_text, help_title = title, is_h3 = TRUE,
  ...)

title_with_tooltip(title, help_text, ...)

title_with_popover(title, help_text, help_title = title, ...)
}
\arguments{
\item{title}{title}

\item{help_text}{help_text}

\item{help_title}{help_title}

\item{is_h3}{a boolean indicating if title should be wrapped in \code{h3}}

\item{...}{additional argument passed on to \code{\link[bsplus]{bs_modal}}}
}
\description{
Title with help text in an informative modal, popover or tooltip
}
