% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enable.R
\name{shinymetrics_enable}
\alias{shinymetrics_enable}
\alias{shinymetrics_disable}
\title{Toggle}
\usage{
shinymetrics_enable(
  ...,
  expires = 400,
  session = shiny::getDefaultReactiveDomain()
)

shinymetrics_disable(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{...}{Passed to Cookie options, \href{https://developer.mozilla.org/en-US/docs/Web/HTTP/Cookies}{MDN}.}

\item{expires}{Number of days from time of creation when the cookie expires.
or an object of class \code{Date}, the date when the cookie should expire,
or an empty or \code{NULL} to make it a session cookie.}

\item{session}{A valid Shiny session.}
}
\description{
Enable and disable the tracking from the server.
}
