% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{trk}
\alias{trk}
\title{Track}
\arguments{
\item{ignore}{\code{inputId}s or \code{outputId}s to ignore
from the tracking.
Must include the namespace.
Cannot be used with \code{only}.}

\item{only}{\code{inputId}s or \code{outputId}s to limit
the tracking the tracking to.
Must include the namespace.
Cannot be used with \code{ignore}.}
}
\description{
Track
}
\keyword{internal}
