% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{user}
\alias{user}
\alias{shinymetrics_set_user}
\alias{shinymetrics_unset_user}
\title{Set and Unset User}
\usage{
shinymetrics_set_user(
  user,
  expires = 9999,
  session = shiny::getDefaultReactiveDomain()
)

shinymetrics_unset_user(
  expires = 9999,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{user}{String identifying the user.}

\item{expires}{Number of days from time of creation when the cookie expires.}

\item{session}{A valid Shiny session.}
}
\description{
Set or unset the user being tracked.
This is set, \strong{in plain text} as cookie.
}
