% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{useShinymetrics}
\alias{useShinymetrics}
\title{Dependency}
\usage{
useShinymetrics()
}
\description{
Shinymetrics dependencies.
If you use the \code{track} method of the
\link{Shinymetrics} class you \strong{do not}
need this function.
It is only needed if you want to manually
set the tracking via JavaScript, if you use
functions of the package this is not needed.
}
