% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_metric}
\alias{preview_metric}
\title{Preview a metric in a shinydashboard}
\usage{
preview_metric(metric, ...)
}
\arguments{
\item{metric}{a \code{tbl_metric} object}

\item{...}{additional parameters passed to \code{metric_panel}}
}
\description{
Preview a metric in a shinydashboard
}
\examples{
library(dplyr)
\dontrun{
metric_satisfaction <- datacampr::tbl_metric_product_survey_avg_satisfaction()
preview_metric({
  metric_satisfaction \%>\%
  filter(period == 'rolling_28d')
})
metric <- datacampr::tbl_metric_content_courses_avg_rating_wtd()
preview_metric({
  metric \%>\%
    filter(period == 'rolling_28d')
})
preview_metric({
  d <- metric \%>\%
    filter(period == 'rolling_28d') \%>\%
    filter(launch_status == "live") \%>\%
    filter(technology == "R")
})
}
}
