% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{shinymetrics_server}
\alias{shinymetrics_server}
\title{Initialise server tracking}
\usage{
shinymetrics_server(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{A valid Shiny session.}
}
\description{
This functions is only required if you want to
make use of \code{\link[=shinymetrics_custom_event]{shinymetrics_custom_event()}}.
}
\note{
You can also observe whether Shinymetrics tracking
is enabled by listening to the \code{input$shinymetricsEnabled} input.
}
